Reorganized, see users, and create new user on database
I reorganized the file structure a bit to make it easier to see which is backend and what is not. I also created two webpages /users and /users/add that display all users, and allows you to add a new user. These are run off the app.js file in src. Everything not in the backend folder is frontend. I tried having them in different folders, but github gave me difficulties. The react app in the "frontend" (everything not in backend folder) is launched via "nodemon app" and outputs webpages to  http://localhost:3000/. you can go to http://localhost:3000/users to see the list of users on database. The backend is created via the express generator, and I simply added placed the database files from the branch Database&Server. The backend runs on http://localhost:4000/. You can run it via "nodemon www". In order to display the users on /users you must have the backend and frontend running at the same time by typing both nodemon commands on two different terminals. I will provide better documentation later.