Authentication
Created a folder middleware that verifies the token the user has matches the token we have in our .env file .  This will only allow users with a token to do certain things on the website. I created tokenIsValid post to check the integrity of the user , it Returns a boolean.  I created a .get route for finding the user with the token they were assigned, and a delete route for deleting an account using the token